cal relative 반환값 double로 변경

+++ diaryProject/DiaryScape/DiaryScape/src/main/java/diaryMap/DiaryScape/web/openApi/NodeController.java
@@ -171,7 +171,7 @@ public class NodeController {
     public static String calRelativeX(String startX, String currX){
         double relativeVal = Double.parseDouble(currX) - Double.parseDouble(startX);
         double mul = 100; //위도상 0.063(5.6km거리)는 500을 곱하여 30을 반환하기로 했다.
-        return Integer.toString((int)(relativeVal * mul) * 5);
+        return Double.toString((Double)(relativeVal * mul) * 5);
     }
